01_introduction_agentic_ai/requirements.txt
browser-use
playwright
python-dotenv